made n unsigned int

+++ 0x04-more_functions_nested_loops/101-print_number.c
@@ -6,7 +6,7 @@
  */
 void print_number(int n)
 {
-	int k = n;
+	unsigned int k = n;
 
 	if (n < 0)
 	{
@@ -20,6 +20,6 @@ void print_number(int n)
 	if (k != 0)
 		print_number(k);
 
-	_putchar(n % 10 + '0');
+	_putchar((unsigned int) n % 10 + '0');
 
 }
